+++ src/main/java/com/dodatabase/demo/domain/member/Member.java
@@ -1,0 +1,18 @@
+package com.dodatabase.demo.domain.member;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Entity
+@Data
+@AllArgsConstructor
+public class Member {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long memId;
+}
