--- src/main/java/com/dodatabase/movie_backend/MovieBackendApplication.java
@@ -1,15 +1,0 @@
-package com.dodatabase.movie_backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-public class MovieBackendApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MovieBackendApplication.class, args);
-    }
-
-}
