--- src/main/java/com/dodatabase/movie_backend/config/ModelMapperConfig.java
@@ -1,23 +1,0 @@
-package com.dodatabase.movie_backend.config;
-
-import org.modelmapper.ModelMapper;
-import org.modelmapper.config.Configuration.AccessLevel;
-import org.modelmapper.convention.MatchingStrategies;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-    private final ModelMapper modelMapper = new ModelMapper();
-
-    @Bean
-    public ModelMapper modelMapper() {
-        modelMapper.getConfiguration()
-                .setMatchingStrategy(MatchingStrategies.STRICT)
-                .setFieldAccessLevel(AccessLevel.PRIVATE)
-                .setFieldMatchingEnabled(true);
-        return modelMapper;
-    }
-
-}
