--- src/main/java/com/dodatabase/movie_backend/domain/Member/Member.java
@@ -1,18 +1,0 @@
-package com.dodatabase.movie_backend.domain.Member;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-import lombok.*;
-
-@Entity
-@Data
-@AllArgsConstructor
-public class Member {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long memId;
-}
