+++ src/test/java/com/dodatabase/demo/DemoApplicationTests.java
@@ -1,0 +1,23 @@
+package com.dodatabase.demo;
+
+import static org.junit.Assert.assertNull;
+
+import org.junit.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
+
+@SpringBootTest
+@AutoConfigureMockMvc
+public class DemoApplicationTests {
+
+  @Autowired
+  private ApplicationContext applicationContext;
+
+  @Test
+  public void contextLoads() {
+    assertNull(applicationContext);
+  }
+
+}
