+++ src/test/java/com/dodatabase/demo/TestConfig.java
@@ -1,0 +1,15 @@
+package com.dodatabase.demo;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+@TestConfiguration
+public class TestConfig {
+
+  @Bean
+  public ModelMapper modelMapper() {
+    return new ModelMapper();
+  }
+
+}
