+++ src/test/java/com/dodatabase/demo/controller/JpaControllerTest.java
@@ -1,0 +1,8 @@
+package com.dodatabase.demo.controller;
+
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+
+@WebMvcTest(JpaController.class)
+public class JpaControllerTest {
+
+}
