+++ src/test/java/com/dodatabase/demo/repository/MovieApiRepositoryTest
@@ -1,0 +1,5 @@
+package com.dodatabase.demo.repository;
+
+public class MovieApiRepositoryTest extends MovieBackendApplicationTests {
+
+}
