--- src/test/java/com/dodatabase/movie_backend/MovieBackendApplicationTests.java
@@ -1,10 +1,0 @@
-package com.dodatabase.movie_backend;
-
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@AutoConfigureMockMvc
-class MovieBackendApplicationTests {
-
-}
