--- src/test/java/com/dodatabase/movie_backend/TestConfig.java
@@ -1,15 +1,0 @@
-package com.dodatabase.movie_backend;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration
-public class TestConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-
-}
