--- src/test/java/com/dodatabase/movie_backend/controller/JpaControllerTest.java
@@ -1,8 +1,0 @@
-package com.dodatabase.movie_backend.controller;
-
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-
-@WebMvcTest(JpaController.class)
-public class JpaControllerTest {
-
-}
