--- src/test/java/com/dodatabase/movie_backend/repository/MovieApiRepositoryTest
@@ -1,5 +1,0 @@
-package com.dodatabase.movie_backend.repository;
-
-public class MovieApiRepositoryTest extends MovieBackendApplicationTests {
-    
-}
